Support unrestricted ports usage for custom URI schemes
Some custom URI schemes may assign a different meaning to the port
of an URI. Webkit restricts, by default, usage of certain ports.
The new API allows configuring which URI schemes are allowed to
use all ports. Network based schemes, such as http, continue to
have the ports restriction in place and bypass is not possible
using the new API.